Create README.md
This program gathers the activity of a Discord user and displays a status icon and text description via a bot which can be employed within a gaming server. Statuses may include 'offline', the name of a program being run on the user's computer, or the name of a game being played within the Steam platform.